add: 地図に関するUIや処理をrecord_mapに移動しました
